Buttons to select trails Implemented
Changes to map.jsx / map.css. Removed the drop down list of trails 1- 4 and replaced with 3 buttons. Buttons can be clicked to highlight the associated trail. Button supports showing trail length and name, but that is not implemented yet.

Slight changes to css to fix fitting of NOAA data icons and text.

All parts of website work as intended before pushing.